fix(store): extract catalog-upgrade allow preservation helper
Move Allow All persistence coverage into a focused test module so oversized
authority files do not grow.
